requirements.txt
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ddgs==9.10.0
fake-useragent==2.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
primp==0.15.0
python-dotenv==1.2.1
requests==2.32.5
socksio==1.0.0
typing_extensions==4.15.0
urllib3==2.6.3
